Linux64 compilation fixes and baseline updates.
SVN: r1322@trunk

+++ src/plugins/SensitivityAnalysis.cpp
@@ -1189,7 +1189,7 @@ namespace mesmer
       int tableSize = int(table.size());
       m_pParallelManager->broadcastInteger(&tableSize, 1, i);
       for (size_t j(0); j < size_t(tableSize); j++) {
-        if (m_rank == i)
+        if (size_t(m_rank) == i)
           tmp = table[j];
         m_pParallelManager->broadcastDouble(&tmp[0], tmp.size(), i);
         wrk.push_back(tmp);
